refactor(approval): split allowance.rs into sub-modules for token efficiency
Split the 1750-line allowance.rs monolith into 6 focused files,
all under 500 lines, to improve agent token efficiency:

- mod.rs (212): AllowanceId, Allowance struct, re-exports, type tests
- pattern.rs (358): AllowancePattern enum, matches(), helpers, Display
- pattern_tests.rs (425): non-plugin pattern matching tests
- pattern_plugin_tests.rs (266): plugin pattern matching tests
- store.rs (255): AllowanceStore implementation
- store_tests.rs (291): store tests

No behavioral changes. All 200 tests pass.
